requirement.txt
langchain_openai
python-box
streamlit
langchain
langchain_community
huggingface_hub
python-dotenv